mongodb2/requirements.txt
pymongo
pandas
pyyaml
mysql-connector-python
pyodbc
